+++ Assets/Scripts/DOTS/Data.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 53fa0fffe9bf8cf48b8a317042b69646
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
