+++ Assets/Scripts/DOTS/Data/TimeToLive.cs
