+++ Assets/Scripts/DOTS/Systems/TimedDestroySystem.cs
