+++ Assets/Sprites/bigAsteroid.png.meta
